requirements.txt
# Core ML and Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
xgboost>=1.7.0
lightgbm>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0

# Configuration and Utilities
pyyaml>=6.0
omegaconf>=2.3.0
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Time Series and Climate
statsmodels>=0.14.0
pmdarima>=2.0.0
darts>=0.24.0
xarray>=2023.5.0
netCDF4>=1.6.0

# Geospatial
geopandas>=0.13.0
shapely>=2.0.0
rasterio>=1.3.0
rioxarray>=0.15.0
pyproj>=3.5.0
folium>=0.14.0
contextily>=1.4.0

# Remote Sensing and Computer Vision
opencv-python>=4.8.0
scikit-image>=0.21.0

# Optimization
cvxpy>=1.3.0

# Monitoring (optional)
mlflow>=2.5.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.25.0

# Development Tools
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0
